requirements.txt
kafka.python
requests
schedule
